flowglow/requirements.txt
gradio==4.19.2
groq==0.4.2
langchain==0.1.0
python-dotenv==1.0.0
pydantic==2.5.2
pytest==7.4.3
black==23.11.0
flake8==6.1.0"
pydantic-settings==2.1.0